+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add isotropic flux source direction
+- Fixed bug in Bins error parsing
 
 ## [0.0.4] - 2021-10-21
 - Implemented free gas scattering